build: Add a meta xorgproto.pc
Currently each protocol has its own version. Also, the standalone
repositories for each of those protocols is deprecated in favor of this
xorgproto package.

Add an xorgproto.pc that contains the actual package version, so that
applications do not have to list each protocol and the version they
require, but can chose to require instead just a particular version of
xorgproto.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>